Implement uploading apk and sending request for energy evaluation
Switched to dropzone npm package for drag and drop ability of apk. On
upload success another HTTP request is sent to do perform the energy
evaluation. For now, it only take text responses, but the plan is to take
the output graph data results from Orka and display them to the user.