Add ability to add and remove multiple app requests
This will be used for batch requests. + and - icons have been
added that will add a new row and delete a row respectively.
Uses Font-Awesome for the icons. The dropzones are now situated
side by side so that it's more clear which monkeyrunner scripts
are tied to which apk upload. Flex-grow is used to make dropzone
fill the width.